use existing math helpers
